Atualização
Ajax e js

+++ veterinario/app/Http/Controllers/ClienteController.php
@@ -13,28 +13,10 @@ class ClienteController extends Controller
         return view('cliente.index', compact('clientes'));
     }
 
-    public function create()
-    {
-        return view('cliente.create');
-    }
+    public function create() {}
 
     public function store(Request $request)
     {
-        $regras = [
-            'nome' => 'required|max:100|min:10',
-            'email' => 'required|unique:clientes',
-            'telefone' => 'required|max:13|min:11',
-        ];
-
-        $msgs = [
-            'required' => 'O preenchimento do campo [:attribute] é obrigatório!',
-            'max' => 'O campo [:attribute] possui tamanho máximo de [:max] caracteres!',
-            'min' => 'O campo [:attribute] possui tamanho mínimo de [:min] caracteres!',
-            'email.unique' => 'Já existe um cliente cadastrado para o e-mail especificado!'
-        ];
-
-        $request->validate($regras, $msgs);
-
         $cliente = new Cliente([
             'nome' => $request->get('nome'),
             'email' => $request->get('email'),
@@ -43,57 +25,51 @@ class ClienteController extends Controller
 
         $cliente->save();
 
-        return redirect()->route('cliente.index');
+        return json_encode($cliente);
     }
 
     public function show($id)
     {
-        $dados = Cliente::find($id);
-        return view('cliente.show', compact('dados'));
-    }
-
-    public function edit($id)
-    {
         $cliente = Cliente::find($id);
-        return view('cliente.edit', compact('cliente'));
 
+        if(isset($cliente)){
+            return json_encode($cliente);
+        }
+
+        return response('Cliente não encontrado', 404);
     }
+
+    public function edit($id){}
 
     public function update(Request $request, $id)
     {
-        $regras = [
-            'nome' => 'required|max:100|min:10',
-            'telefone' => 'required|max:13|min:11',
-        ];
-
-        $msgs = [
-            'required' => 'O preenchimento do campo [:attribute] é obrigatório!',
-            'max' => 'O campo [:attribute] possui tamanho máximo de [:max] caracteres!',
-            'min' => 'O campo [:attribute] possui tamanho mínimo de [:min] caracteres!',
-        ];
-
-        $request->validate($regras, $msgs);
-
-        $cliente = new Cliente([
-            'nome' => $request->get('nome'),
-            'email' => $request->get('email'),
-            'telefone' => $request->get('telefone')
-        ]);
-
         $cliente = Cliente::find($id);
-        $cliente->nome = $request->get('nome');
-        $cliente->email = $request->get('email');
-        $cliente->telefone = $request->get('telefone');
-        $cliente->save();
 
-        return redirect()->route('cliente.index');
+        if(isset($cliente)){
+            $cliente->nome = $request->get('nome');
+            $cliente->email = $request->get('email');
+            $cliente->telefone = $request->get('telefone');
+            $cliente->save();
+            return json_encode($cliente);
+        }
+
+        return response('Cliente não encontrado', 404);
     }
 
     public function destroy($id)
     {
         $cliente = Cliente::find($id);
-        $cliente->delete();
 
-        return redirect()->route('cliente.index');
+        if(isset($cliente)){
+            $cliente->delete();
+            return response('OK', 200);
+        }
+
+        return response('Cliente não encontrado', 404);
+    }
+
+    public function loadJson(){
+        $cliente = Cliente::all();
+        return json_encode($cliente);
     }
 }
